Revert "Attempt ungeneric interface"
This reverts commit 12a300d17a8f8044dced5bc31525634c89d8404b.
